Templates
Task 1
Write a template function that performs a linear search
for a user-defined value, displays its index on the screen. The array of values is not
ordered, the size is set randomly (from 30 to 100). Array Values
also set randomly (up to 100).

Task 2
Create a parameterized binary tree class. With methods - add an element
to the tree, traversing the tree in descending and ascending order.
Perform a tree search.
The tree consists of nodes that