Inheritance
Create a parent class "Queue" with the functions of initializing the queue,
adding an item to the queue and retrieving an item from the queue.
Create a method for creating a copy of the queue. The result should be a new instance
of the “Queue” class consisting of elements (copies of elements) of the original queue. The order
of the elements must be preserved.
Create a function for merging two queues. The result should be a queue
consisting of elements of the first queue and the second queue. The order of the elements
must be preserved.
Based on the parent class "Queue", create a child class "Queue1" with
the functions of finding and displaying on the screen: Finding the last element smaller than the arithmetic mean