Operator overloading
• Task 1 Unary operation
Create a "Date" class that contains a date, month, and year. Overload
operators: ++ (prefix form, method) increases either the date,
month, or year by 1 at the user's choice. -- (prefix form, method) similarly decreases.
++ (postfix form, friendly function) increases either the
date, month, or year by a user number at the user's choice. -- (postfix form, method)
similarly reduces.
• Task 2. Binary operation
Create a queue type object. Overload the + operator as a member function and * as
a friendly function. + adds an item to the queue, * multiplies the items in the queue
by a custom number. Pulls an element from the queue with the - operator. Queues can
be assigned (=), checked for equality (element-wise) == or !=, input and output to the
stream.